// src/lib/type.ts
export type Content = {
	type:
	| 'paragraph'
	| 'quote'
	| 'heading_2'
	| 'heading_3';
	text: string | null;
} | {
	type: 'code';
	text: string | null;
	language: string | null;
};


export type Post = {
	id: string;
	title: string | null;
	slug: string | null;
	createdTs: string | null;
	lastEditedTs: string | null;
	contents: Content[];
};